product, chatのリファクタリング
