+++ src/features/product/components/templates/EditProduct.tsx
@@ -8,16 +8,15 @@ import { PathToProductPage } from "@/features/product/types/product";
 import { useRouter } from "next/router";
 import { useState } from "react";
 import { useForm } from "react-hook-form";
-import { useRecoilValue } from "recoil";
 import { ProductFormField } from "@/features/product/components/molecules/Fierld/ProductFormField";
 import { Loading } from "../../../../components/organisms/Loading/Loading";
 import { ProductLikeButton } from "@/features/product/components/molecules/Button/ProductLikeButton";
-import { EmployeeState } from "@/stores/employeeAtom";
-import { UserCard } from "@/features/user/components/organisms/Card/UserCard";
 import { useProductWithApprovedUserRecruitParticipants } from "../../hooks/useProductWithApprovedUserRecruitParticipants";
 import { ProductFormSkillsField } from "../molecules/Fierld/ProductFormSkillsField";
 import { SkillSelect } from "@/components/molecules/Select/SkillSelect";
 import { useIsRelatedUserByRecruitId } from "../../hooks/useIsRelatedUserByRecruitId";
+import { OurOwnCommentsList } from "../organisms/List/OurOwnCommentsList";
+import { RelatedUsersList } from "../organisms/List/RelatedUsersList";
 
 type Props = {
   path: PathToProductPage;
@@ -27,8 +26,6 @@ export const EditProduct = ({ path }: Props): JSX.Element => {
   const router = useRouter();
   const { id } = router.query;
 
-  const user = useRecoilValue(UserState);
-  const employee = useRecoilValue(EmployeeState);
   const { product } = useProductWithApprovedUserRecruitParticipants(
     id as string
   );
@@ -40,7 +37,7 @@ export const EditProduct = ({ path }: Props): JSX.Element => {
     control,
     formState: { errors },
   } = useForm();
-  const [isModalOpen, setIsModalOpen] = useState(false);
+
   const [isNameOpen, setIsNameOpen] = useState(false);
   const [isSkillsOpen, setIsSkillsOpen] = useState(false);
   const [isReasonForSkillSelectionOpen, setIsReasonForSkillSelectionOpen] =
@@ -48,25 +45,15 @@ export const EditProduct = ({ path }: Props): JSX.Element => {
   const [isDevelopmentBackgroundOpen, setIsDevelopmentBackgroundOpen] =
     useState(false);
   const [isOverviewOpen, setIsOverviewOpen] = useState(false);
-  const [isCommentOpen, setIsCommentOpen] = useState(false);
 
-  const closeModal = () => {
-    setIsModalOpen(false);
-  };
-
-  //いいねをしているかしていないかの判定に利用する
-  const isLiked = product?.employeeToProductLikes?.some(
-    (like) => like.employeeId === employee?.id
-  );
-
-  if (product === undefined || isLiked === undefined) return <Loading />;
+  if (!product) return <Loading />;
 
   return (
     <div className="flex flex-col w-full min-h-screen h-full justify-center items-center text-gray-600 ">
       <div className="flex flex-col rounded-lg items-center w-4/5 sm:w-sm md:w-md lg:w-lg gap-14  bg-white">
         <div className="flex justify-center items-center w-full mt-10">
           <video
-            src={product?.url}
+            src={product.url}
             controls
             className="w-full h-auto rounded-sm"
           ></video>
@@ -75,7 +62,7 @@ export const EditProduct = ({ path }: Props): JSX.Element => {
           {/* headline */}
           <ProductFormField
             labelText="プロダクト名"
-            input={product?.name}
+            input={product.name}
             editable={isRelatedUser}
             onCLick={() => setIsNameOpen(true)}
           />
@@ -96,7 +83,7 @@ export const EditProduct = ({ path }: Props): JSX.Element => {
           {/* skills */}
           <ProductFormSkillsField
             labelText="スキル"
-            skills={product?.skills}
+            skills={product.skills}
             onCLick={() => setIsSkillsOpen(true)}
           />
           <EditProductModal
@@ -177,105 +164,26 @@ export const EditProduct = ({ path }: Props): JSX.Element => {
             />
           </EditProductModal>
 
-          {/* Fix: 分岐しているコードを見やすくする */}
-          {/* プロダクト参加者のアイコンを一覧表示したい。それぞれのユーザーのプロフィールに飛ぶことができる */}
-          <div className="flex flex-col items-start w-full gap-5 my-10 border-t border-gray-400 pt-5">
-            <div className="flex flex-row justify-between w-full">
-              <div className="font-semibold p-2 text-left">
-                個人アピールポイント一覧
-              </div>
-              {/* userの時のみコメントを作成できる */}
-              {path === PathToProductPage.UserPath && (
-                <div>
-                  {!product?.comment?.some(
-                    (comment) => comment.userId === user?.id
-                  ) ? (
-                    <button
-                      onClick={() => setIsModalOpen(true)}
-                      className="py-2 px-6 rounded-md text-white font-bold bg-green-500 hover:bg-green-600"
-                    >
-                      Comment作成
-                    </button>
-                  ) : (
-                    product?.comment.map((comment) => (
-                      <div key={comment.id}>
-                        {comment.userId === user?.id && <p>作成済み</p>}
-                      </div>
-                    ))
-                  )}
-                </div>
-              )}
-            </div>
-            {!product?.comment && (
-              <div className="p-2 rounded-md text-white font-bold bg-red-500">
-                ※アピールポイントを追加してください
-              </div>
-            )}
-            {product?.comment?.map((comment, index) => {
-              return (
-                <div key={index} className="flex flex-col w-full">
-                  {user?.id === comment.userId ? (
-                    <>
-                      <ProductFormField
-                        labelText={comment.user.name}
-                        input={comment.content}
-                        onCLick={() => setIsCommentOpen(true)}
-                      />
-                      <EditCommentModal
-                        isOpen={isCommentOpen}
-                        setIsOpen={setIsCommentOpen}
-                        commentId={comment.id}
-                        handleSubmit={handleSubmit}
-                      >
-                        <PlainTextArea
-                          registerLabel="content"
-                          labelText="アピールポイントを記述してください。"
-                          placeholder="FEの分野で〇〇に取り組みました....."
-                          register={register}
-                        />
-                      </EditCommentModal>
-                    </>
-                  ) : (
-                    <div className="w-full flex flex-col">
-                      <div className="flex flex-row w-full mb-3">
-                        <p className="font-semibold w-1/2">
-                          {comment.user.name}
-                        </p>
-                      </div>
-                      <div className="border rounded-md p-4 border-gray-300">
-                        {comment.content}
-                      </div>
-                    </div>
-                  )}
-                </div>
-              );
-            })}
-          </div>
+          {/* コメント一覧 */}
+          <OurOwnCommentsList 
+            path={path}
+            comments={product.comment!}
+            productId={product?.id}
+          />
 
           {/* 関わった学生ユーザー一覧 */}
-          <div className="w-full flex gap-4 flex-wrap">
-            {product.approvedUserRecruitParticipants.map((participant) => (
-              <div key={participant.id}>
-                <UserCard user={participant.user} />
-              </div>
-            ))}
-          </div>
+          <RelatedUsersList
+            participantsInfo={product.approvedUserRecruitParticipants}
+          />
 
+          {/* 企業のみ押すことができるいいねボタン */}
           {path === PathToProductPage.CorporationPath && (
             <ProductLikeButton
-              productId={product.id as string}
-              isPropsLiked={isLiked}
+              product={product}
             />
           )}
         </div>
       </div>
-
-      <AddCommentModal
-        isOpen={isModalOpen}
-        closeModal={closeModal}
-        productId={product?.id!}
-        setIsOpen={setIsModalOpen}
-      />
     </div>
   );
 };
